abstract more functions to utils
